requirements.txt
litellm==1.83.4
python-dotenv==1.2.2
